Move Check to models
