+++ uptime/models.py
@@ -2,6 +2,8 @@ import os
 import logging
 import socket
 import uuid
+import json
+import datetime
 
 logging.getLogger('uptime')
 
@@ -62,3 +64,45 @@ class Config:
                 )
             )
             setattr(self, current_name, new_value)  # Override the config
+
+
+class Check(object):
+    """
+    URL Check configuration object
+    """
+
+    def __init__(self, check_json):
+        defaults = {'content': None,
+                    'interval': 15}
+
+        self.url = None
+        self.__dict__ = json.loads(check_json)
+
+        # use defaults for undefined optional params
+        for k, v in defaults.items():
+            if k not in self.__dict__:
+                self.__setattr__(k, v)
+
+        self.check_id = str(self.check_id)
+        self.name = self.check_id
+
+        self.failures = 0
+        self.notified = False
+
+        self.last = datetime.datetime.utcnow()
+
+        self.interval = int(self.interval)
+        logging.info('loaded check %s for url %s' % (self.check_id, self.url))
+
+    def dump_json(self):
+        #  ret = json.clone(self.__dict__)
+        ret = copy.copy(self.__dict__)
+        del ret['last']
+        return json.dumps(ret)
+
+    def ok(self):
+        """
+        Method to reset failures and notify switches following successful check
+        """
+        self.failures = 0
+        self.notified = False
